Adds Posthog analytics integration
Integrates Posthog analytics to track user behavior and gather insights.

This change adds a Posthog component and includes it in the base head of the web application. This enables tracking of user interactions, page views, and other relevant data for improved understanding of user behavior.

Also configures `.gitignore` to ignore Vercel files and local environment files.